requirements.txt
Click==7.0
cycler==0.10.0
dill==0.2.9
dominate==2.4.0
Flask==1.0.3
Flask-Bootstrap==3.3.7.1
inspyred==1.0.1
itsdangerous==1.1.0
Jinja2==2.10.1
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.0.3
numpy==1.16.4
pkg-resources==0.0.0
pyKriging==0.2.0
pymssql==2.1.4
pyparsing==2.4.0
python-dateutil==2.8.0
python-dotenv==0.10.3
scipy==1.3.0
six==1.12.0
visitor==0.1.3
Werkzeug==0.15.4
